internal: move validation of number of checktypes URLs to config
Calling checktype.NewCatalog with no URLs should be valid. In that
case, the returned catalog will be empty.

At the same time, we want to force users to provide at least one
catalog URL in the configuration file. Otherwise, no security checks
would be run, which is not an obvious behavior. Thus, this commit
moves this validation to the config package, which also validates that
at least one target is provided.