internal/engine: make target server ignore unsupported asset types
Lava currently reports a warning message when a target cannot be
handled by the engine's target server:

	2024-04-28T11:55:40+02:00 WARN could not handle target
	target={alpine:latest DockerImage map[]} err=generate stream:
	get target addr: invalid asset type: DockerImage

In some cases, for instance when the target is a DockerImage, this is
expected. However, it can be misleading for users that do not know the
Lava internal details. Thus, this commit makes the target server
explicitly handle supported asset types.